services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
influxdb-client==1.38.0
